Split utilities and types into separate files to improve readability
